src/main/java/com/ticod/timesailerbackend/util/EncryptionUtil.java
package com.ticod.timesailerbackend.util;

import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.security.SecureRandom;
import java.util.Base64;

public class EncryptionUtil {

    private static final int SALT_SIZE = 16;
    private static final String SHA_256 = "SHA-256";

    public static String generateSalt() {
        SecureRandom secureRandom = new SecureRandom();
        byte[] salt = new byte[SALT_SIZE];
        secureRandom.nextBytes(salt);
        return Base64.getEncoder().encodeToString(salt);
    }

    public static String generateHash(String target, String salt) throws NoSuchAlgorithmException {
        MessageDigest md = MessageDigest.getInstance(SHA_256);
        md.update((target + salt).getBytes());
        return Base64.getEncoder().encodeToString(md.digest());
    }

}
